+++ slider/src/commands/addressCommand.h
@@ -1,0 +1,27 @@
+#ifndef ADDRESSCOMMAND_H
+#define ADDRESSCOMMAND_H
+
+#include "src/core/menu/menuSystem.h"
+#include "src/app/brain/connectApp.h"
+#include <memory>
+
+using namespace Core;
+
+namespace Slider
+{
+    class BrainAddressCommand : public MenuCommand
+    {
+    public:
+        virtual void Print(Core::Display* display) const override;
+        virtual void Invoke(MenuCommandButton command) override {}
+    };
+
+    class ControllerAddressCommand : public MenuCommand
+    {
+    public:
+        virtual void Print(Core::Display* display) const override;
+        virtual void Invoke(MenuCommandButton command) override {}
+    };
+}
+
+#endif
